+++ tests/integration/test_hostgroups.py
@@ -352,9 +352,8 @@ def test_remove_owner_nonexistent(
 
     assert msg == snapshot(
         """\
-mreg_api.exceptions.DeleteError: 404 Not Found
-  DELETE <server-url>/api/v1/hostgroups/<hostgroup-name>/owners/exampleuser1
-  Owner 'exampleuser1' is not associated with host group '<hostgroup-name>'.\
+mreg_api.exceptions.DeleteError: 404 Not Found: DELETE <server-url>/api/v1/hostgroups/<hostgroup-name>/owners/exampleuser1
+Owner 'exampleuser1' is not associated with host group '<hostgroup-name>'.\
 """
     )